roadmap-agent/requirements.txt
fastapi
uvicorn
langchain_community
langchain
openai
matplotlib
plotly
python-dotenv
#pinecone
boto3
schedule
langchain-google-genai
beautifulsoup4
isodate
pydantic
sqlalchemy
psycopg2-binary
python-jose
passlib
bcrypt
python-multipart